PCHR-2150: Add abstract class for drupal interaction forms
