+++ uk.co.compucorp.civicrm.hrcore/CRM/HRCore/Form/CreateUserRecordTaskForm.php
@@ -1,40 +1,14 @@
 <?php
 
 use CRM_Utils_Array as ArrayHelper;
-use CRM_HRCore_Service_DrupalUserService as DrupalUserService;
+use CRM_HRCore_Form_AbstractDrupalInteractionTaskForm as AbstractDrupalInteractionTaskForm;
 
-class CRM_HRCore_Form_CreateUserRecordTaskForm extends CRM_Contact_Form_Task {
-
-  /**
-   * @var array
-   */
-  protected $contactDetails = [];
+class CRM_HRCore_Form_CreateUserRecordTaskForm extends AbstractDrupalInteractionTaskForm {
 
   /**
    * @var bool
    */
   protected $sendEmail = FALSE;
-
-  /**
-   * @var DrupalUserService
-   */
-  protected $drupalUserService;
-
-  /**
-   * @param null $state
-   * @param mixed $action
-   * @param string $method
-   * @param null $name
-   */
-  public function __construct(
-    $state = NULL,
-    $action = CRM_Core_Action::NONE,
-    $method = 'post',
-    $name = NULL
-  ) {
-    $this->drupalUserService = Civi::container()->get('drupal_user_service');
-    parent::__construct($state, $action, $method, $name);
-  }
 
   /**
    * @inheritdoc
@@ -50,7 +24,6 @@ class CRM_HRCore_Form_CreateUserRecordTaskForm extends CRM_Contact_Form_Task {
    */
   public function preProcess() {
     parent::preProcess();
-    $this->initContactDetails();
     $haveNoAccount = $this->getContactsWithoutAttribute('uf_id');
     $haveAccount = array_diff_key($this->contactDetails, $haveNoAccount);
     $this->assign('contactsWithoutEmail', $this->getContactsWithoutAttribute('email'));
@@ -109,71 +82,6 @@ class CRM_HRCore_Form_CreateUserRecordTaskForm extends CRM_Contact_Form_Task {
     }
 
     return $user;
-  }
-
-  /**
-   * Creates an array to store contact details
-   */
-  private function initContactDetails() {
-    $emailParams = [
-      'contact_id' => '$value.id',
-      'return' => ['email'],
-      'is_primary' => 1,
-    ];
-    $ufMatchParams = [
-      'contact_id' => '$value.id',
-      'return' => ['uf_id']
-    ];
-    $params = [
-      'return' => ['display_name'],
-      'id' => ['IN' => $this->_contactIds],
-      'options' => ['limit' => 0],
-      'api.Email.getsingle' => $emailParams,
-      'api.UFMatch.getsingle' => $ufMatchParams,
-    ];
-    $contactDetails = civicrm_api3('Contact', 'get', $params);
-    $contactDetails = ArrayHelper::value('values', $contactDetails, []);
-
-    foreach ($contactDetails as $detail) {
-      $contactID = (int) ArrayHelper::value('contact_id', $detail);
-      $displayName = ArrayHelper::value('display_name', $detail);
-      $ufMatch = ArrayHelper::value('api.UFMatch.getsingle', $detail, []);
-      $ufId = (int) ArrayHelper::value('uf_id', $ufMatch);
-      $emailDetails = ArrayHelper::value('api.Email.getsingle', $detail, []);
-      $email = ArrayHelper::value('email', $emailDetails);
-
-      $this->setContactDetail($contactID, 'display_name', $displayName);
-      $this->setContactDetail($contactID, 'uf_id', $ufId);
-      $this->setContactDetail($contactID, 'email', $email);
-      $this->setContactDetail($contactID, 'id', $contactID);
-    }
-  }
-
-  /**
-   * Helper method to prevent overwriting of contact details
-   *
-   * @param int $contactId
-   * @param string $type
-   * @param mixed $value
-   */
-  private function setContactDetail($contactId, $type, $value) {
-    if (!isset($this->contactDetails[$contactId][$type])) {
-      $this->contactDetails[$contactId][$type] = $value;
-    }
-  }
-
-  /**
-   * @param string $property
-   *  The property to check if empty
-   *
-   * @return array
-   */
-  protected function getContactsWithoutAttribute($property) {
-    $checker = function ($contactDetail) use ($property) {
-      return empty($contactDetail[$property]);
-    };
-
-    return array_filter($this->contactDetails, $checker);
   }
 
   /**
